reviewed JSONata processors
